Replace VMDriver.DiskPath with ImageFormat
DiskPath is currently provided by the code using the machine, but
it must match the path the actual machine driver will copy its disk image to

This is quite fragile, and the path/name the VM image was copied to can
be kept as an internal implementation detail to the  machine driver.
This commit adds a 'VMDriver.ImageFormat' field though, as this is useful in
order to generate the filename the source image will be copied to.