Add UpdateConfigRaw method to drivers
This will be useful to dynamically change some driver parameters such as memory
size, or disk size.

This needs to be called with a serialized driver structure containing new
values for the parameters we want to update. Machine drivers will then check
these parameters with their current config, and try to adjust.